stock_alert/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
django==4.1.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
idna==3.4
    # via requests
pytz==2023.3
    # via djangorestframework
requests==2.30.0
    # via -r requirements.in
sqlparse==0.4.4
    # via django
urllib3==2.0.2
    # via requests
